move data streams messages code
