+++ comp/core/autodiscovery/providers/datastreams/kafka_messages_test.go
